requirements.txt
python_vlc==3.0.20123
Requests==2.32.3
urllib3==1.26.14
